requirements.txt
flask
pandas
scikit-learn
sentence-transformers
faiss-cpu
seaborn
matplotlib